strict-rules の指摘 22 件を解消
いずれも挙動を変えない修正。

- empty($array) を $array === [] に変更（Arr::first / Arr::last）
- 短縮三項演算子を明示的な条件式に展開（Application / Str / テストの getenv）
- 不要な型キャストを削除（array_rand の (int)、getallheaders の (array)）
- !preg_match(...) を preg_match(...) !== 1 に変更（int|false の否定を避ける）
- Result::count() の PHPDoc を int<0, max> に変更（Countable との共変性）
- テストの getenv() ?: をヘルパー envOr() に集約

+++ src/Sloop/Database/Replica/RandomReplicaSelector.php
@@ -24,6 +24,6 @@ final class RandomReplicaSelector implements ReplicaSelector
      */
     public function pick(array $candidates): int
     {
-        return (int) array_rand($candidates);
+        return array_rand($candidates);
     }
 }
